Fixed wrong filenames of unit tests + New str func + tests
